refactor(blorktools): asset-panel directory
