+++ packages/blorktools/src/asset_debugger/asset-panel/asset-panel.html
@@ -1,5 +1,5 @@
 <!-- Asset Panel Component -->
-<link rel="stylesheet" href="../../css/asset-panel.css">
+<link rel="stylesheet" href="/asset_debugger/asset-panel/asset-panel.css">
 <div id="asset-tab" class="tab-content">
     <div class="asset-section">
         <div class="collapsible-header section-header">
