+++ packages/blorktools/src/asset_debugger/ui/scripts/settings-modal.js
@@ -1,7 +1,7 @@
 /**
  * Settings Modal UI Component for Asset Debugger
  */
-import { saveSettings, loadSettings, getDefaultSettings } from '../../data/localstorage-util.js';
+import { saveSettings, loadSettings, getDefaultSettings } from '../../util/localstorage-util.js';
 import { updateRigOptions } from '../../rig/rig-manager.js';
 
 export class SettingsModal {
